@@ -6,47 +6,47 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 0.0.2
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  redux: ^4.0.0
-  redux_epics: ^0.14.0
-  flutter_redux: ^0.6.0
-  built_value: ^6.0.0
-  built_collection: ^4.0.0
-  http: ^0.12.0
-  json_annotation: ^3.0.0
-  oauth2: ^1.6.0
-  flutter_web_auth: ^0.2.2
-  shared_preferences: ^0.5.7+1
-  cupertino_icons: ^0.1.3
-  flutter_slidable: ^0.5.4
-  path_to_regexp: ^0.3.0
-  quiver: ^2.1.3
-  receive_sharing_intent: ^1.4.0+2
-  carousel_slider: ^2.0.0
-  reselect: ^0.4.0
-  easy_debounce: ^1.0.1
-  amplitude_flutter: ^2.0.0
-  assets_audio_player:
-    git:
-      url: https://github.com/florent37/Flutter-AssetsAudioPlayer.git
-      ref: b49a4360f6a7f182d4a6e23ebcdbf43ad570913b
-  sentry_flutter: ^4.0.4
+  http: ^1.1.0
+  json_annotation: ^4.8.1
+  oauth2: ^2.0.2
+  flutter_web_auth: ^0.5.0
+  shared_preferences: ^2.2.1
+  font_awesome_flutter: '>= 4.7.0'
+  carousel_slider: ^4.2.1
+  easy_debounce: ^2.0.3
+  amplitude_flutter: ^3.16.0
+  sentry_flutter: ^7.9.0
+  drift: ^2.11.1
+  sqlite3_flutter_libs: ^0.5.0
+  path_provider: ^2.0.0
+  path: ^1.8.3
+  provider: ^6.1.1
+  freezed_annotation: ^2.4.1
 
 dev_dependencies:
-  build_runner: ^1.0.0
-  built_value_generator: ^6.0.0
-  json_serializable: ^3.0.0
+  build_runner: ^2.4.7
+  built_value_generator: ^8.6.2
+  json_serializable: ^6.7.1
   flutter_test:
     sdk: flutter
+  drift_dev: ^2.11.1
+  freezed: ^2.4.5
 
 flutter:
   uses-material-design: true
   assets:
     - assets/btn_google_signin_dark_normal_web@2x.png
+  fonts:
+    - family: Raleway
+      fonts:
+        - asset: fonts/Raleway/Raleway-Regular.ttf
+        - asset: fonts/Raleway/Raleway-Italic.ttf
+          style: italic
 
   # To add assets to your application, add an assets section, like this:
   # assets:
