@@ -3,16 +3,16 @@ import 'package:vockify/src/api/dto/translate/translation_dto.dart';
 
 part 'dictionary_entry_dto.g.dart';
 
-@JsonSerializable(nullable: false)
+@JsonSerializable(createToJson: false)
 class DictionaryEntryDto {
   @JsonKey(name: 'text')
-  final String text;
+  final String? text;
 
   @JsonKey(name: 'pos')
-  final String partOfSpeech;
+  final String? partOfSpeech;
 
   @JsonKey(name: 'ts')
-  final String transcription;
+  final String? transcription;
 
   @JsonKey(name: 'tr')
   final List<TranslationDto> translations;
@@ -20,6 +20,4 @@ class DictionaryEntryDto {
   DictionaryEntryDto(this.text, this.partOfSpeech, this.transcription, this.translations);
 
   factory DictionaryEntryDto.fromJson(Map<String, dynamic> json) => _$DictionaryEntryDtoFromJson(json);
-
-  Map<String, dynamic> toJson() => _$DictionaryEntryDtoToJson(this);
 }
