@@ -20,7 +20,7 @@ enum NavigationType {
 
 /// The action to be dispatched in the store in order to trigger a navigation.
 class NavigateToAction {
-  final String name;
+  final String? name;
 
   /// Controls the method to be called on the [Navigator] with the specified
   /// [name].
@@ -28,25 +28,27 @@ class NavigateToAction {
 
   /// Optional callback function to be called before the actual navigation.
   /// e.g. activate the loader.
-  final Function preNavigation;
+  final Function? preNavigation;
 
   /// Optional callback function to be called after the actual navigation.
   /// e.g. de-activate the loader.
-  final Function postNavigation;
+  final Function? postNavigation;
 
   /// Optional object to be passed either in [NavigationType.shouldPush] or
   /// [NavigationType.shouldReplace].
   ///
   /// It will be ignored if passed with type [NavigationType.shouldPop]
-  final Object arguments;
+  final Object? arguments;
 
   /// Optional object to be passed either in [NavigationType.shouldPushNamedAndRemoveUntil]
   /// or in [NavigationType.shouldPopUntil]
   ///
   /// It will be ignored if passed with any other type.
-  final RoutePredicate predicate;
+  final RoutePredicate? predicate;
 
-  final VoidCallback onPop;
+  final VoidCallback? onPop;
+
+  final NavigatorState? currentState;
 
   /// Create a navigation action.
   ///
@@ -58,7 +60,8 @@ class NavigateToAction {
       this.postNavigation,
       this.arguments,
       this.predicate,
-      this.onPop})
+      this.onPop,
+      this.currentState})
       : assert(() {
           if (type == NavigationType.shouldPushNamedAndRemoveUntil) {
             return predicate != null;
@@ -73,27 +76,26 @@ class NavigateToAction {
         }());
 
   factory NavigateToAction.push(String name,
-          {Function preNavigation,
-          Function postNavigation,
-          Object arguments,
-          VoidCallback onPop}) =>
+          {Function? preNavigation,
+          Function? postNavigation,
+          Object? arguments,
+          VoidCallback? onPop,
+          NavigatorState? currentState}) =>
       NavigateToAction(name,
           preNavigation: preNavigation,
           postNavigation: postNavigation,
           arguments: arguments,
-          onPop: onPop);
+          onPop: onPop,
+          currentState: currentState);
 
-  factory NavigateToAction.pop(
-          {Function preNavigation, Function postNavigation}) =>
+  factory NavigateToAction.pop({Function? preNavigation, Function? postNavigation, NavigatorState? currentState}) =>
       NavigateToAction(null,
           type: NavigationType.shouldPop,
           preNavigation: preNavigation,
-          postNavigation: postNavigation);
+          postNavigation: postNavigation,
+          currentState: currentState);
 
-  factory NavigateToAction.popUntil(
-          {Function preNavigation,
-          Function postNavigation,
-          RoutePredicate predicate}) =>
+  factory NavigateToAction.popUntil({Function? preNavigation, Function? postNavigation, RoutePredicate? predicate}) =>
       NavigateToAction(null,
           type: NavigationType.shouldPopUntil,
           preNavigation: preNavigation,
@@ -101,28 +103,31 @@ class NavigateToAction {
           predicate: predicate);
 
   factory NavigateToAction.replace(String name,
-          {Function preNavigation,
-          Function postNavigation,
-          Object arguments,
-          VoidCallback onPop}) =>
+          {Function? preNavigation,
+          Function? postNavigation,
+          Object? arguments,
+          VoidCallback? onPop,
+          NavigatorState? currentState}) =>
       NavigateToAction(name,
           type: NavigationType.shouldReplace,
           preNavigation: preNavigation,
           postNavigation: postNavigation,
           arguments: arguments,
-          onPop: onPop);
+          onPop: onPop,
+          currentState: currentState);
 
-  factory NavigateToAction.pushNamedAndRemoveUntil(
-          String name, RoutePredicate predicate,
-          {Function preNavigation,
-          Function postNavigation,
-          Object arguments,
-          VoidCallback onPop}) =>
+  factory NavigateToAction.pushNamedAndRemoveUntil(String name, RoutePredicate predicate,
+          {Function? preNavigation,
+          Function? postNavigation,
+          Object? arguments,
+          VoidCallback? onPop,
+          NavigatorState? currentState}) =>
       NavigateToAction(name,
           type: NavigationType.shouldPushNamedAndRemoveUntil,
           preNavigation: preNavigation,
           postNavigation: postNavigation,
           predicate: predicate,
           arguments: arguments,
-          onPop: onPop);
+          onPop: onPop,
+          currentState: currentState);
 }
