@@ -6,23 +6,25 @@ class PrimaryButtonWidget extends StatelessWidget {
   final VoidCallback onPressed;
 
   const PrimaryButtonWidget({
-    Key key,
-    @required this.text,
-    @required this.onPressed,
+    Key? key,
+    required this.text,
+    required this.onPressed,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return FlatButton(
-      color: VockifyColors.fulvous,
+    return TextButton(
       onPressed: onPressed,
-      padding: EdgeInsets.symmetric(
-        vertical: 15,
-        horizontal: 20,
+      style: TextButton.styleFrom(
+        padding: EdgeInsets.symmetric(
+          vertical: 15,
+          horizontal: 20,
+        ),
+        backgroundColor: VockifyColors.fulvous,
       ),
       child: Text(
         text,
-        style: Theme.of(context).textTheme.bodyText1.copyWith(
+        style: Theme.of(context).textTheme.bodyLarge?.copyWith(
               color: VockifyColors.white,
               fontSize: 16,
             ),
