@@ -1,17 +1,17 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:vockify/src/theme/vockify_colors.dart';
 
 class PrimaryTextFormFieldWidget extends StatelessWidget {
   final TextEditingController controller;
   final String label;
-  final Widget suffix;
+  final Widget? suffix;
   final bool autoFocus;
   final bool isRequired;
 
   const PrimaryTextFormFieldWidget({
-    Key key,
-    @required this.controller,
-    @required this.label,
+    Key? key,
+    required this.controller,
+    required this.label,
     this.suffix,
     this.autoFocus = false,
     this.isRequired = true,
@@ -22,19 +22,31 @@ class PrimaryTextFormFieldWidget extends StatelessWidget {
     return Padding(
       padding: EdgeInsets.only(bottom: 16),
       child: TextFormField(
+        style: TextStyle(
+          color: VockifyColors.prussianBlue,
+          fontSize: 20,
+          fontWeight: FontWeight.bold,
+        ),
         autofocus: autoFocus,
         controller: controller,
         decoration: InputDecoration(
+          floatingLabelBehavior: FloatingLabelBehavior.always,
           suffixIcon: suffix,
-          labelText: label,
+          label: Text(
+            label.toUpperCase(),
+            style: TextStyle(
+              color: VockifyColors.prussianBlue,
+              fontSize: 18,
+              fontWeight: FontWeight.normal,
+            ),
+          ),
           fillColor: Colors.white,
-          border: OutlineInputBorder(
+          border: UnderlineInputBorder(
             borderSide: BorderSide(),
-            borderRadius: BorderRadius.all(Radius.circular(2.0)),
           ),
         ),
         validator: (value) {
-          if (isRequired && value.isEmpty) {
+          if (isRequired && (value?.isEmpty ?? false)) {
             return 'ОБЯЗАТЕЛЬНОЕ ПОЛЕ';
           }
 
