@@ -6,15 +6,9 @@ part of 'translate_response.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-TranslateResponse _$TranslateResponseFromJson(Map<String, dynamic> json) {
-  return TranslateResponse(
-    (json['data'] as List)
-        .map((e) => DictionaryEntryDto.fromJson(e as Map<String, dynamic>))
-        .toList(),
-  );
-}
-
-Map<String, dynamic> _$TranslateResponseToJson(TranslateResponse instance) =>
-    <String, dynamic>{
-      'data': instance.data,
-    };
+TranslateResponse _$TranslateResponseFromJson(Map<String, dynamic> json) =>
+    TranslateResponse(
+      (json['def'] as List<dynamic>)
+          .map((e) => DictionaryEntryDto.fromJson(e as Map<String, dynamic>))
+          .toList(),
+    );
