@@ -6,11 +6,10 @@ part of 'translation_dto.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-TranslationDto _$TranslationDtoFromJson(Map<String, dynamic> json) {
-  return TranslationDto(
-    json['text'] as String,
-  );
-}
+TranslationDto _$TranslationDtoFromJson(Map<String, dynamic> json) =>
+    TranslationDto(
+      json['text'] as String,
+    );
 
 Map<String, dynamic> _$TranslationDtoToJson(TranslationDto instance) =>
     <String, dynamic>{
