@@ -1,17 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:vockify/src/router/routes.dart';
-import 'package:vockify/src/screens/flashcards_screen.dart';
-import 'package:vockify/src/screens/login_screen.dart';
-import 'package:vockify/src/screens/main_screen.dart';
-import 'package:vockify/src/screens/profile_screen.dart';
-import 'package:vockify/src/screens/public_terms_screen.dart';
-import 'package:vockify/src/screens/quiz_screen.dart';
-import 'package:vockify/src/screens/set_screen.dart';
-import 'package:vockify/src/screens/start_screen.dart';
-import 'package:vockify/src/screens/tour_screen.dart';
-import 'package:vockify/src/screens/user_set_select_screen.dart';
-import 'package:vockify/src/screens/user_term_screen.dart';
-import 'package:vockify/src/screens/user_terms_screen.dart';
+import 'package:vockify/src/pages/flashcards/flashcards_page.dart';
+import 'package:vockify/src/pages/sets/sets_page.dart';
+import 'package:vockify/src/pages/terms/public_terms_page.dart';
+import 'package:vockify/src/pages/quiz/quiz_page.dart';
+import 'package:vockify/src/pages/set/set_page.dart';
+import 'package:vockify/src/pages/start/start_page.dart';
+import 'package:vockify/src/pages/tour/tour_page.dart';
+import 'package:vockify/src/pages/set_select/user_set_select_page.dart';
+import 'package:vockify/src/pages/term/user_term_page.dart';
+import 'package:vockify/src/pages/terms/user_terms_page.dart';
 import 'package:vockify/src/services/amplitude.dart';
 import 'package:vockify/src/widgets/home.dart';
 
@@ -19,28 +17,26 @@ typedef Widget PathBuilder(Map<String, dynamic> arguments);
 
 class AppRouter {
   static final Map<String, PathBuilder> _paths = {
-    Routes.login: (arguments) => LoginScreenWidget(),
-    Routes.profile: (arguments) => ProfileScreenWidget(),
-    Routes.tour: (arguments) => TourScreenWidget(),
-    Routes.start: (arguments) => StartScreenWidget(),
-    Routes.userSetSelect: (arguments) => UserSetSelectScreenWidget(
+    Routes.tour: (arguments) => TourPageWidget(),
+    Routes.start: (arguments) => StartPageWidget(),
+    Routes.userSetSelect: (arguments) => UserSetSelectPageWidget(
           term: arguments['term'] as String,
           definition: arguments['definition'] as String,
         ),
-    Routes.main: (arguments) => MainScreenWidget(),
-    Routes.home: (arguments) => HomeWidget(intent: arguments['intent'] as String),
-    Routes.publicTerms: (arguments) => PublicTermsScreenWidget(setId: arguments['id'] as int),
-    Routes.userSet: (arguments) => SetScreenWidget(setId: arguments['id'] as int),
-    Routes.quiz: (arguments) => QuizScreenWidget(setId: arguments['setId'] as int),
-    Routes.flashcards: (arguments) => FlashcardsScreenWidget(setId: arguments['setId'] as int),
-    Routes.userTerms: (arguments) => UserTermsScreenWidget(setId: arguments['id'] as int),
-    Routes.userTerm: (arguments) => UserTermScreenWidget(
+    Routes.main: (arguments) => SetsPageWidget(),
+    Routes.home: (arguments) => HomeWidget(),
+    Routes.publicTerms: (arguments) => PublicTermsPageWidget(setId: arguments['id'] as int),
+    Routes.userSet: (arguments) => SetPageWidget(setId: arguments['id'] as int?),
+    Routes.quiz: (arguments) => QuizPageWidget(setId: arguments['setId'] as int),
+    Routes.flashcards: (arguments) => FlashcardsPageWidget(setId: arguments['setId'] as int),
+    Routes.userTerms: (arguments) => UserTermsPageWidget(setId: arguments['id'] as int),
+    Routes.userTerm: (arguments) => UserTermPageWidget(
           setId: arguments['setId'] as int,
-          termId: arguments['termId'] as int,
+          termId: arguments['termId'] as int?,
         ),
   };
 
-  static MaterialPageRoute buildRoute(RouteSettings settings, Widget builder) {
+  static MaterialPageRoute<dynamic> buildRoute(RouteSettings settings, Widget builder) {
     amplitude.logEvent(
       'route_changed',
       eventProperties: {'name': settings.name, 'arguments': settings.arguments.toString()},
@@ -52,13 +48,23 @@ class AppRouter {
     );
   }
 
-  static Route getRoute(RouteSettings settings) {
+  static Route<dynamic> getRoute(RouteSettings settings) {
     final arguments = settings.arguments;
 
     if (arguments != null && arguments is! Map<String, dynamic>) {
       throw ArgumentError("arguments cast error");
     }
 
-    return buildRoute(settings, _paths[settings.name](arguments as Map<String, dynamic>));
+    if (arguments == null) {
+
+    }
+
+    final pathBuilder = _paths[settings.name];
+
+    if (pathBuilder != null) {
+      return buildRoute(settings, pathBuilder(arguments as Map<String, dynamic>? ?? Map<String, dynamic>()));
+    }
+
+    throw Exception('Couldn\'t find path name');
   }
 }
