@@ -2,20 +2,18 @@ class Routes {
   // app routes
   static const String app = "/";
   static const String tour = "/tour";
-  static const String home = "/home/:intent";
-  static const String login = "/login";
+  static const String home = "/home";
   static const String quiz = "/quiz/:setId";
   static const String flashcards = "/flashcards/:setId";
 
   // home routes
   static const String start = "/start";
   static const String main = "/main";
-  static const String profile = "/profile";
 
   // user set routes
   static const String userSet = "/user_set";
   static const String userSetSelect = '/user_set_select';
-  
+
   // user term routes
   static const String userTerms = "/user_terms";
   static const String userTerm = "/user_term";
