@@ -1,7 +1,7 @@
 class AuthorizationException implements Exception {
   final String message;
 
-  const AuthorizationException([this.message]);
+  const AuthorizationException(this.message);
 
   String toString() => 'AuthorizationException: $message';
 }
